BC-68: Monaco Menu infront of othe fields in Form
Adjust z-index for editor container and styling

Updated the z-index of the main container from 1 to 0 to modify its stacking order. Added a new CSS rule for the `.editor` class to set its z-index to 1 when it contains a direct child of the `.monaco-editor-container`, ensuring proper layering of the editor.